+++ frontend-react/src/components/GalleryList.js
@@ -1,6 +1,9 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
 import { Link } from "react-router-dom";
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faEdit, faTrashAlt } from '@fortawesome/free-solid-svg-icons';
+import Swal from 'sweetalert2';
 
 const GalleryList = () => {
   const [galleries, setGalleries] = useState([]);
@@ -15,12 +18,35 @@ const GalleryList = () => {
   };
 
   const deleteGallery = async (galleryId) => {
-    try {
-      await axios.delete(`http://localhost:5000/galleries/${galleryId}`);
-      getGalleries();
-    } catch (error) {
-      console.log(error);
-    }
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        axios.delete(`http://localhost:5000/galleries/${galleryId}`)
+          .then(() => {
+            Swal.fire(
+              'Deleted!',
+              'Your gallery has been deleted.',
+              'success'
+            );
+            getGalleries();
+          })
+          .catch(error => {
+            console.error('Error deleting the gallery:', error);
+            Swal.fire(
+              'Error!',
+              'Failed to delete the gallery.',
+              'error'
+            );
+          });
+      }
+    });
   };
 
   return (
@@ -41,8 +67,12 @@ const GalleryList = () => {
               <td>{gallery.name}</td>
               <td>{gallery.description}</td>
               <td>
-                <Link to={`/editgallery/${gallery.id}`}>Edit</Link>
-                <button onClick={() => deleteGallery(gallery.id)}>Delete</button>
+                <Link to={`/editgallery/${gallery.id}`}>
+                  <FontAwesomeIcon icon={faEdit} />
+                </Link>
+                <button onClick={() => deleteGallery(gallery.id)}>
+                  <FontAwesomeIcon icon={faTrashAlt} />
+                </button>
               </td>
             </tr>
           ))}
